Replaced Collection View Controller with a Table View Controller
